+++ src/newswatch/scrapers/suara.py
@@ -25,8 +25,8 @@ from .basescraper import BaseScraper
 
 
 class SuaraScraper(BaseScraper):
-    def __init__(self, keywords, concurrency=12, start_date=None, queue_=None):
-        super().__init__(keywords, concurrency, queue_)
+    def __init__(self, keywords, concurrency=12, start_date=None, queue_=None, keyword_concurrency=None):
+        super().__init__(keywords, concurrency, queue_, keyword_concurrency=keyword_concurrency)
         self.base_url = "https://www.suara.com"
         self.start_date = start_date
         self.continue_scraping = True